+++ src/app.html
@@ -5,35 +5,6 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<meta name="theme-color" content="#00ffee" />
 		<meta name="msapplication-TileColor" content="#213253" />
-		<meta name="title" content="bDesigned" />
-		<meta name="description" content="Svelte Sirens is a Svelte Society for women & non-binary." />
-
-		<!-- Open Graph / Facebook -->
-		<meta property="og:type" content="website" />
-		<meta property="og:url" content="https://www.sveltesirens.dev" />
-		<meta property="og:title" content="Svelte Sirens" />
-		<meta
-			property="og:description"
-			content="Svelte Sirens is a Svelte Society for women & non-binary."
-		/>
-		<meta
-			property="og:image"
-			content="https://www.sveltesirens.dev/images/android-chrome-192x192.png"
-		/>
-
-		<!-- Twitter -->
-		<meta property="twitter:card" content="summary" />
-		<meta property="twitter:url" content="https://twitter.com/BrittneyPostma" />
-		<meta property="twitter:title" content="Svelte Sirens" />
-		<meta
-			property="twitter:description"
-			content="Svelte Sirens is a Svelte Society for women & non-binary."
-		/>
-		<meta
-			property="twitter:image"
-			content="https://www.sveltesirens.dev/images/android-chrome-192x192.png"
-		/>
-		<meta name="twitter:image:alt" content="Svelte Sirens Logo" />
 
 		<link rel="manifest" href="/site.webmanifest" crossorigin="use-credentials" />
 		<link rel="apple-touch-icon" sizes="180x180" href="/images/apple-touch-icon.png" />
