+++ src/routes/__layout.svelte
@@ -1,31 +1,12 @@
 <script>
 	import '@styles/global.scss';
-	import { page } from '$app/stores';
 
+	import Head from '@layout/head.svelte';
 	import Bubbles from '@components/bubbles.svelte';
 	import Header from '@layout/header.svelte';
-
-	const capitalize = (str, lower = false) =>
-		(lower ? str.toLowerCase() : str).replace(/(?:^|\s|["'([{])+\S/g, (match) =>
-			match.toUpperCase()
-		);
-	let title =
-		$page.path === '/'
-			? 'Svelte Sirens'
-			: `${$page.path.substring(1).replace('s/', ' | ').replace(/-/g, ' ')} | Svelte Sirens`;
 </script>
 
-<svelte:head>
-	<link rel="icon" type="image/svg+xml" href="/images/favicon.svg" />
-	<link rel="icon" type="image/png" href="/images/favicon.png" />
-	<link rel="preconnect" href="https://fonts.googleapis.com" />
-	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-	<link
-		href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative:wght@400;700&family=Red+Hat+Text:wght@300;400;700&display=swap"
-		rel="stylesheet"
-	/>
-	<title>{capitalize(title)}</title>
-</svelte:head>
+<Head />
 
 <Header />
 <main>
